Multiple bundles, optimise rollup config, coreJS to resolve regeneratorRuntime issue.
