+++ package.json
@@ -41,6 +41,7 @@
     "@rollup/plugin-babel": "^5.3.0",
     "@rollup/plugin-commonjs": "^17.1.0",
     "@rollup/plugin-node-resolve": "^11.2.0",
+    "core-js": "^3.9.1",
     "cpy": "8.1.1",
     "cross-env": "7.0.2",
     "cssnano": "^4.1.10",
