+++ rollup.config.js
@@ -1,44 +1,75 @@
 import { nodeResolve as resolve } from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
-import { getBabelOutputPlugin } from '@rollup/plugin-babel';
+import babel from '@rollup/plugin-babel';
 import { terser } from 'rollup-plugin-terser';
 
 const env = process.env.NODE_ENV;
 
-export default {
-  input: 'src/js/main.js',
-
-  watch: {
-    include: 'src/js',
-    clearScreen: false,
-  },
-
-  plugins: [
-    getBabelOutputPlugin({
-      presets: ['@babel/preset-env'],
-    }),
-    resolve({
-      browser: true,
-    }),
-    commonjs(),
-  ],
-
-  output: {
-    file: 'dist/js/bundle.js',
-    format: 'es',
-    plugins: [
-      terser({
-        ecma: 2018,
-        mangle: { toplevel: true },
-        compress: {
-          module: true,
-          toplevel: true,
-          unsafe_arrows: true,
-          drop_console: !env,
-          drop_debugger: !env,
+const babelConfig = {
+  babelHelpers: 'bundled',
+  exclude: 'node_modules/**',
+  presets: [
+    [
+      '@babel/preset-env',
+      {
+        targets: {
+          browsers: ['ie 11'],
         },
-        output: { quote_style: 1 },
-      }),
+        useBuiltIns: 'usage',
+        corejs: 3,
+      },
     ],
-  },
+  ],
 };
+
+const terserConfig = {
+  ecma: 2018,
+  mangle: { toplevel: true },
+  compress: {
+    module: true,
+    toplevel: true,
+    unsafe_arrows: true,
+    drop_console: !env,
+    drop_debugger: !env,
+  },
+  output: { quote_style: 1 },
+};
+
+export default [
+  // Main bundle used on every page
+  {
+    input: 'src/js/main.js',
+
+    plugins: [
+      commonjs(),
+      resolve({
+        browser: true,
+      }),
+      babel(babelConfig),
+    ],
+
+    output: {
+      file: 'dist/js/bundle-main.js',
+      format: 'iife',
+      plugins: [terser(terserConfig)],
+    },
+  },
+  // Search bundle used only for blog search
+  {
+    input: 'src/js/search.js',
+
+    plugins: [
+      commonjs(),
+      resolve({
+        browser: true,
+      }),
+      babel(babelConfig),
+    ],
+
+    output: {
+      file: 'dist/js/bundle-search.js',
+      format: 'iife',
+      plugins: [terser(terserConfig)],
+    },
+  },
+];
