+++ src/_includes/layouts/_base.njk
@@ -50,6 +50,6 @@
     </main>
 
     {% include 'components/site-footer.njk' %}
-    <script src="/js/bundle.js"></script>
+    <script src="/js/bundle-main.js"></script>
   </body>
 </html>
