+++ src/_includes/layouts/listing-blog-search.njk
@@ -13,5 +13,4 @@ permalink: '/{{ locale }}/news/blog/search/'
 
 <div id="search-results"></div>
 
-<script src="https://unpkg.com/lunr/lunr.js"></script>
-<script src="/js/search.js"></script>
+<script src="/js/bundle-search.js"></script>
