+++ src/js/main.js
@@ -1,8 +1,8 @@
+import 'details-polyfill';
+import 'focus-visible';
 import { NavSecondary } from './nav-secondary';
 import { OffCanvas } from './off-canvas';
 import { StickyScroll } from './sticky-scroll';
-import 'details-polyfill';
-import 'focus-visible';
 
 NavSecondary.init();
 OffCanvas.init();
